requirements.txt
flask
scikit-learn
pandas
pickle-mixin
pymongo[srv]
certifi
python-dotenv
